requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
distlib==0.3.4
dnspython==2.1.0
filelock==3.4.2
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
is-disposable-email==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
pipenv==2021.11.23
platformdirs==2.4.1
plotly==5.4.0
psycopg2-binary==2.9.2
pycparser==2.21
pygal==3.0.0
python-dateutil==2.8.2
python-decouple==3.5
python-editor==1.0.4
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
tenacity==8.0.1
urllib3==1.26.7
virtualenv==20.11.2
virtualenv-clone==0.5.7
Werkzeug==2.0.2
wikipedia==1.4.0
WTForms==3.0.0
